feat : 좋아요 API, 장바구니 API 수정
기존에는 req.body에서 받아왔던 userId를 header에서 받은 토큰을 decode 하여 사용하도록 변경

+++ week_8-1/controller/LikeController.js
@@ -1,11 +1,27 @@
 const conn = require("../mariadb");
 const { StatusCodes } = require("http-status-codes");
+const jwt = require('jsonwebtoken');
+const dotenv = require('dotenv');
+dotenv.config();
 
 const addLike = (req, res) => {
     const {bookId} = req.params;
-    const {userId} = req.body;
+    const authorization = ensureAuthorization(req);
+
+    if(authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    }
+
+    if(authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    }
+
     const sql = `INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)`;
-    const values = [userId, bookId];
+    const values = [authorization.id, bookId];
     conn.query(sql, values, function(err, results) {
         if(err) {
             console.log(err);
@@ -17,9 +33,22 @@ const addLike = (req, res) => {
 
 const removeLike = (req, res) => {
     const {bookId} = req.params;
-    const {userId} = req.body;
+    const authorization = ensureAuthorization(req);
+
+    if(authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    }
+
+    if(authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    }
+
     const sql = `DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?`;
-    const values = [userId, bookId];
+    const values = [authorization.id, bookId];
     conn.query(sql, values, function(err, results) {
         if(err) {
             console.log(err);
@@ -27,6 +56,18 @@ const removeLike = (req, res) => {
         }
         return res.status(StatusCodes.OK).json(results);
     });
+}
+
+function ensureAuthorization(req, res) {
+    try {
+        const receivedJwt = req.headers["authorization"];
+        const decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
+        return decodedJwt;
+    } catch(err) {
+        console.log(err.name);
+        console.log(err.message);
+        return err;
+    }
 }
 
 module.exports = {
